% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkValue.R
\name{all_binary}
\alias{all_binary}
\title{All binary?}
\usage{
all_binary(x, logical_ok = TRUE)
}
\arguments{
\item{x}{The data vector or matrix}

\item{logical_ok}{Is a logical vector or matrix also acceptable? Default:
\code{TRUE}.}
}
\value{
Logical. Is \code{x} binary data?
}
\description{
Check if a data vector or matrix is all zeroes and ones. Option to also
accept logical values.
}
