% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FC.R
\name{color_palette}
\alias{color_palette}
\title{Color palette}
\usage{
color_palette(pal = "Beach")
}
\arguments{
\item{pal}{\code{"Beach"} (default; blue to white to red),
\code{"Sand"} (white to red), or
\code{"Water"} (white to blue).}
}
\value{
A data.frame with two columns: \code{"col"} has the hex code of color,
and \code{"val"} has the placement of colors between zero and one.
}
\description{
Color palettes for fMRI data analysis tasks
}
