% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCor.R
\name{CompCor.noise_comps}
\alias{CompCor.noise_comps}
\title{CompCor: get noise components}
\usage{
CompCor.noise_comps(X_noise, center, scale, noise_nPC)
}
\arguments{
\item{X_noise}{The noise ROIs data}

\item{center, scale}{Center & scale robustly}

\item{noise_nPC}{Number of PCs to obtain for each noise ROI}
}
\value{
A list with components X, X_noise, ROI_data, ROI_noise, noise_nPC,
noise_erosion, noise_comps, and noise_var.
}
\description{
Get noise components for aCompCor.
}
\keyword{internal}
