% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_volume.R
\name{crop_vol}
\alias{crop_vol}
\title{Crop a 3D array}
\usage{
crop_vol(x)
}
\arguments{
\item{x}{The numeric 3D array to crop.}
}
\value{
A list of length two: \code{"data"}, the cropped array, and
\code{"padding"}, the number of slices removed from each edge of each
dimension.
}
\description{
Remove empty (zero-valued) edge slices from a 3D array.
}
