% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3ddespike.R
\name{despike_3D.interpolate}
\alias{despike_3D.interpolate}
\title{despike_3D from AFNI step 2: interpolate}
\usage{
despike_3D.interpolate(qreg, c1 = 2.5, c2 = 4)
}
\arguments{
\item{qreg}{the quantile regression from \code{\link{despike_3D.qreg}}}

\item{c1}{spike threshold. Default: \code{2.5}.}

\item{c2}{upper range of the acceptable deviation from the fit. Default:
\code{4}.}
}
\description{
Identify and interpolate the outliers for \code{\link{despike_3D}}.
}
\keyword{internal}
