% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3ddespike.R
\name{despike_3D.original}
\alias{despike_3D.original}
\title{Temporary}
\usage{
despike_3D.original(Yt, c1 = 2.5, c2 = 4)
}
\arguments{
\item{Yt, c1, c2}{temporary}
}
\description{
Temporary
}
\keyword{internal}
