% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/despike_3D.R
\name{despike_3D.qreg}
\alias{despike_3D.qreg}
\title{3dDespike from AFNI, step 1}
\usage{
despike_3D.qreg(Yt)
}
\arguments{
\item{Yt}{The data vector.}
}
\value{
the quantile regression
}
\description{
Compute the quantile regression that \code{\link{despike_3D}} is based on.
}
\keyword{internal}
