% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_coef.R
\name{dice_coef}
\alias{dice_coef}
\title{Dice coefficient}
\usage{
dice_coef(X, Y = NULL)
}
\arguments{
\item{X, Y}{Numeric vectors, or matrices with locations along the rows and
networks along the columns. Dice coefficients will be calculated between
each pair of networks between \code{X} and \code{Y}; if \code{Y} is not
provided, \code{X} will be used for \code{Y}.}
}
\value{
The Dice coefficients in a \eqn{Q_x \times Q_y} matrix, where
\eqn{Q_x} is the number of columns in \code{X}, \eqn{Q_y} is the number of
columns in \code{Y}, and element \eqn{ij} is the Dice coefficient between
network \eqn{i} of \code{X} and network \eqn{j} of \code{Y}.
}
\description{
Dice coefficient
}
