% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dice_overlap.R
\name{dice_overlap}
\alias{dice_overlap}
\title{Dice overlap}
\usage{
dice_overlap(X, Y = NULL)
}
\arguments{
\item{X, Y}{If \code{Y} is \code{NULL}, \code{X} should be a binary matrix,
e.g. with locations along the rows and networks along the columns.

If both are provided, they both must be binary vectors (2 network case).}
}
\value{
The Dice overlap.
}
\description{
Dice overlap
}
