% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_RPs.R
\name{expand_RPs}
\alias{expand_RPs}
\title{Expand realignment parameters (RPs)}
\usage{
expand_RPs(RPs)
}
\arguments{
\item{RPs}{A \eqn{T \times N} numeric matrix, where \eqn{T} is the number of
timepoints and \eqn{N} is the number of RPs (typically six) to expand.}
}
\value{
A \eqn{T \times 4N} numeric matrix, with the first \eqn{N} columns
being the original \code{RPs}, the next \eqn{N} being the differences,
the next \eqn{N} being the squares, and the last \eqn{N} being the
squared differences.
}
\description{
Compute the squares, differences, and square differences of each RP
timeseries.
}
