% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize_DR_oneBOLD}
\alias{harmonize_DR_oneBOLD}
\title{DR step for harmonize}
\usage{
harmonize_DR_oneBOLD(
  BOLD,
  format = c("CIFTI", "xifti", "GIFTI", "gifti", "GIFTI2", "gifti2", "NIFTI", "nifti",
    "RDS", "data"),
  GICA,
  mask = NULL,
  gii_hemi = NULL,
  scale = c("local", "global", "none"),
  scale_sm_surfL = NULL,
  scale_sm_surfR = NULL,
  scale_sm_FWHM = 2,
  TR = NULL,
  hpf = 0.01,
  GSR = FALSE,
  NA_limit = 0.1,
  brainstructures = c("all"),
  varTol = 1e-06,
  maskTol = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{BOLD}{Vector of subject-level fMRI data in one of the following
formats: CIFTI file paths, \code{"xifti"} objects, GIFTI file paths,
\code{"gifti"} objects, NIFTI file paths, \code{"nifti"} objects,
or \eqn{V \times T} numeric matrices, where \eqn{V} is the number of data
locations and \eqn{T} is the number of timepoints.
@param gii_hemi Which hemisphere, in the case of a single GIFTI file?}

\item{format}{Expected format of \code{BOLD}. Should be one
of the following: a \code{"CIFTI"} file path, a \code{"xifti"} object, a
\code{"NIFTI"} file path, a \code{"nifti"} object, or a \code{"data"} matrix.}

\item{GICA}{Group ICA maps in a format compatible with \code{BOLD}. Can also
be a (vectorized) numeric matrix (\eqn{V \times Q}) no matter the format of
\code{BOLD}. Its columns will be centered.}

\item{mask}{Required if the entries of \code{BOLD} are NIFTI
file paths or \code{"nifti"} objects, optional for other formats. For NIFTI, this is a brain map formatted as a
binary array of the same spatial dimensions as the fMRI data, with
\code{TRUE} corresponding to in-mask voxels. For other formats, a logical vector.}

\item{scale}{\code{"global"} (default), \code{"local"}, or \code{"none"}.
Global scaling will divide the entire data matrix by the mean image standard
deviation (\code{mean(sqrt(rowVars(BOLD)))}). Local scaling will divide each
data location's time series by its estimated standard deviation.}

\item{scale_sm_surfL, scale_sm_surfR, scale_sm_FWHM}{Only applies if
\code{scale=="local"} and \code{BOLD} represents surface data (CIFTI or
GIFTI). To smooth the standard deviation estimates used for local scaling,
provide the surface geometries along which to smooth as GIFTI geometry files
or \code{"surf"} objects, as well as the smoothing FWHM (default: \code{2}).

If \code{scale_sm_FWHM==0}, no smoothing of the local standard deviation
estimates will be performed.

If \code{scale_sm_FWHM>0} but \code{scale_sm_surfL} and
\code{scale_sm_surfR} are not provided, the default inflated surfaces from
the HCP will be used.

To create a \code{"surf"} object from data, see
\code{\link[ciftiTools]{make_surf}}. The surfaces must be in the same
resolution as the \code{BOLD} data.}

\item{TR}{The temporal resolution of the data, i.e. the time between volumes,
in seconds. \code{TR} is required for detrending with \code{hpf}.}

\item{hpf}{The frequency at which to apply a highpass filter to the data
during pre-processing, in Hertz. Default: \code{0.01} Hertz. Set to \code{0}
to disable the highpass filter.

The highpass filter serves to detrend the data, since low-frequency
variance is associated with noise. Highpass filtering is accomplished by
nuisance regression of discrete cosine transform (DCT) bases.

Note the \code{TR} argument is required for highpass filtering. If
\code{TR} is not provided, \code{hpf} will be ignored.}

\item{GSR}{Center BOLD across columns (each image)? This
is equivalent to performing global signal regression. Default:
\code{FALSE}.}

\item{brainstructures}{Only applies if the entries of \code{BOLD} are CIFTI
file paths. This is a character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("all")}.}

\item{varTol}{Tolerance for variance of each data location. For each scan,
locations which do not meet this threshold are masked out of the
harmonization.
Default: \code{1e-6}. Variance is calculated on the original data, before
any normalization.}

\item{maskTol}{For computing the dual regression result for each subject:
tolerance for number of locations masked out due to low
variance or missing values. If more than this many locations are masked out,
a subject is not incldued in the harmonization analysis. \code{maskTol}
can be specified either as a proportion of the number of locations (between
zero and one), or as a number of locations (integers greater than one).
Default: \code{.1}, i.e. up to 10 percent of locations can be masked out.}

\item{missingTol}{For harmonizing all subjects:
tolerance for number of subjects masked out due to low variance or missing
values at a given location. If more than this many subjects are masked out,
the location will not be included in the harmonization analysis. \code{missingTol}
can be specified either as a proportion of the number of locations (between
zero and one), or as a number of locations (integers greater than one).
Default: \code{.1}, i.e. up to 10 percent of subjects can be masked out
at a given location.}
}
\description{
Compute dual regression for harmonize
}
\keyword{internal}
