% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hat_matrix.R
\name{hat_matrix}
\alias{hat_matrix}
\title{Hat matrix}
\usage{
hat_matrix(design)
}
\arguments{
\item{design}{The \eqn{T} by \eqn{Q} design matrix}
}
\value{
The \eqn{T} by \eqn{T} hat matrix
}
\description{
Get the hat matrix from a design matrix.
}
\details{
Uses the QR decomposition.
}
\examples{
hat_matrix(cbind(seq(100), 1))
}
