% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FC.R
\name{image.scale}
\alias{image.scale}
\title{image.scale}
\usage{
image.scale(
  z,
  zlim,
  col = color_palette("Beach"),
  breaks,
  axis.pos = 1,
  add.axis = TRUE,
  ...
)
}
\arguments{
\item{z, zlim, col, breaks, axis.pos, add.axis, ...}{The arguments.}
}
\value{
Plots the image scale.
}
\description{
image.scale. Source: r-bloggers.com/2013/12/new-version-of-image-scale-function/
}
\keyword{internal}
