% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkValue.R
\name{is_1}
\alias{is_1}
\title{Is this object the expected data type, and length one?}
\usage{
is_1(x, dtype = c("numeric", "logical", "character"))
}
\arguments{
\item{x}{The value to check}

\item{dtype}{The data type. Default: \code{"numeric"}. Also can be
\code{"logical"} or \code{"character"}}
}
\value{
\code{TRUE} if \code{x} is \code{dtype} and length one.
}
\description{
Is this object the expected data type, and length one?
}
