% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkValue.R
\name{is_constant}
\alias{is_constant}
\title{Is this numeric vector constant?}
\usage{
is_constant(x, TOL = 1e-08)
}
\arguments{
\item{x}{The numeric vector}

\item{TOL}{minimum range of \code{x} to be considered non-constant.
Default: \code{1e-8}}
}
\value{
Is \code{x} constant?
}
\description{
Is this numeric vector constant?
}
