% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_datatype.R
\name{is_integer}
\alias{is_integer}
\title{Is this an integer?}
\usage{
is_integer(x, nneg = FALSE)
}
\arguments{
\item{x}{The putative integer}

\item{nneg}{Require \code{x>=0} (non-negative) too?}
}
\value{
Logical indicating whether \code{x} is an integer
}
\description{
Is this an integer?
}
\keyword{internal}
