% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checkValue.R
\name{is_posNum}
\alias{is_posNum}
\title{Is this object a positive number?}
\usage{
is_posNum(x)
}
\arguments{
\item{x}{The value to check}
}
\value{
\code{TRUE} if \code{x} is a single positive number
}
\description{
Is this object a positive number?
}
\keyword{internal}
