% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R, R/utils_toSort.R
\name{make_mask}
\alias{make_mask}
\title{Create a mask based on vertices that are invalid}
\usage{
make_mask(BOLD, meanTol = -Inf, varTol = 1e-06, verbose = TRUE)

make_mask(BOLD, meanTol = -Inf, varTol = 1e-06, verbose = TRUE)
}
\arguments{
\item{BOLD}{A \eqn{V \times T} numeric matrix. Each row is a location.}

\item{meanTol, varTol}{Tolerance for mean and variance of each data location.
Locations which do not meet these thresholds are masked out of the analysis.
Defaults: \code{-Inf} for \code{meanTol} (ignore), and \code{1e-6} for
{varTol}.}

\item{verbose}{Print messages counting how many locations are removed?}

\item{data}{A list of sessions, where each session is a list with elements
BOLD, design and nuisance. See \code{?create.session} and \code{?is.session}
for more details.}
}
\value{
A logical vector indicating valid vertices

A logical vector indicating valid vertices
}
\description{
Create a mask based on vertices that are invalid

Create a mask based on vertices that are invalid
}
\keyword{internal}
