% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UT.R
\name{mat2UT}
\alias{mat2UT}
\title{Matrix to Upper Triangular Vector}
\usage{
mat2UT(x, diag = FALSE)
}
\arguments{
\item{x}{A square matrix}

\item{diag}{Get the diagonal values too? Default: \code{FALSE}}
}
\value{
The vectorized upper triangle of x.
}
\description{
Returns the vectorized upper triangle of a square matrix
}
