% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_nets.R
\name{match_nets}
\alias{match_nets}
\title{Find best matches for each network based on Dice overlap}
\usage{
match_nets(A, B = "Yeo_7", B_zero_drop = TRUE, brainstructures = "existing")
}
\arguments{
\item{A}{The networks to find matches for. Can be a CIFTI file path, a
\code{"xifti"} object, a numeric matrix (locations by networks, and can be
continuous or binary), or a locations-length vector giving the network index
each location belongs to.}

\item{B}{The networks to match to. Can be a CIFTI file path, a
\code{"xifti"} object, a numeric matrix (locations by networks, and can be
continuous or binary), a locations-length vector giving the network index
each location belongs to, or the name of a parcellation included in
\code{ciftiTools}. Default: \code{"Yeo_7"}.}

\item{B_zero_drop}{If B is a parcellation (i.e. a vector of network indices),
and the lowest parcel index is 0, drop that level? Sometimes the zero-valued
parcel is used for the medial wall or other invalid locations. Default:
\code{TRUE}. If \code{FALSE}, treat it as a candidate for matching just like
all the other parcels.}

\item{brainstructures}{The brainstructures to use, if \code{A} and/or
\code{B} are CIFTI data. Default: \code{"existing"} to use the
brainstructures present in \code{A} (or in \code{B} if \code{A} is a matrix).}
}
\value{
A numeric vector giving the indices of \code{B} networks that are the best
matches for each network in \code{A}. \code{order(match_nets(A, B))} would give an
indexing of \code{A} that rearranges its networks according to the order of
networks in \code{B}.
}
\description{
For each network (map or parcel) in A, it finds the network in B with which
it has the highest Dice coefficient. (Some networks in B will have zero
matches, and some will have more than one.)
}
\details{
If the resolutions of A and B differ, it will try to resample B to match
the resolution of A.
}
