% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_misc.R
\name{Mode}
\alias{Mode}
\title{Mode of data vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{The data vector}
}
\value{
The mode
}
\description{
Get mode of a data vector. But use the median instead of the mode if all
data values are unique.
}
