% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{neighbor_counts}
\alias{neighbor_counts}
\title{Count each voxel's neighbors with value \code{TRUE}}
\usage{
neighbor_counts(arr, pad = FALSE)
}
\arguments{
\item{arr}{The 3D logical array.}

\item{pad}{Pad value for edge.}
}
\value{
An array with the same dimensions as \code{arr}. Each voxel value
will be the number of its immediate neighbors (0 to 6) which are \code{TRUE}.
}
\description{
For each voxel in a 3D logical array, count the number of immediate neighbors
with value \code{TRUE}.
}
\keyword{internal}
