% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{read_gifti_expect_hemi}
\alias{read_gifti_expect_hemi}
\title{Read GIFTI with expected hemisphere}
\usage{
read_gifti_expect_hemi(gii, hemi = c("left", "right"))
}
\arguments{
\item{gii}{A file path to a GIFTI metric file, or a \code{"gifti"} object.}

\item{hemi}{Expected hemisphere. Default: \code{"left"}.}
}
\description{
Read in a GIFTI file, and check that the hemisphere is as expected.
}
\keyword{internal}
