% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{read_gifti_expect_mask}
\alias{read_gifti_expect_mask}
\title{Read GIFTI with mask data}
\usage{
read_gifti_expect_mask(gii, hemi = c("left", "right"))
}
\arguments{
\item{gii}{A file path to a GIFTI metric file, or a \code{"gifti"} object,
with a single column of zeroes and ones.}

\item{hemi}{Expected hemisphere. Default: \code{"left"}.}
}
\description{
Read in a GIFTI file, and check that it has one binary/logical column.
}
\keyword{internal}
