% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nifti.R
\name{read_nifti}
\alias{read_nifti}
\title{Wrapper to common functions for reading NIFTIs}
\usage{
read_nifti(nifti_fname)
}
\arguments{
\item{nifti_fname}{The file name of the NIFTI.}
}
\value{
The NIFTI
}
\description{
Tries \code{RNifti::readNifti}, then \code{oro.nifti::readNIfTI}. If
neither package is available an error is raised.
}
\keyword{internal}
