% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mat.R
\name{rowMedians2}
\alias{rowMedians2}
\title{Row medians}
\usage{
rowMedians2(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{The data matrix}

\item{na.rm}{Ignore NA values? Default: \code{FALSE}}
}
\value{
The row medians of \code{x}.
}
\description{
Use \code{robustbase::rowMedians} package if available, \code{apply} if not.
}
\keyword{internal}
