% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{tangent_space_projection}
\alias{tangent_space_projection}
\title{Tangent space projection}
\usage{
tangent_space_projection(A, B, reverse = FALSE)
}
\arguments{
\item{A, B, reverse}{To-Do}
}
\value{
To-Do
}
\description{
Tangent space projection
}
\keyword{internal}
