% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_design_matrix.R
\name{validate_design_mat}
\alias{validate_design_mat}
\title{Validate design matrix}
\usage{
validate_design_mat(design, T_ = NULL)
}
\arguments{
\item{design}{The design matrix}

\item{T_}{the expected number of rows in \code{design}. Default:
\code{NULL} (no expected value to validate).}
}
\value{
The (modified) design matrix
}
\description{
Coerces \code{design} to a numeric matrix, and optionally checks that the
number of rows is as expected. Sets constant-valued columns to 1, and scales
all other columns.
}
