% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_design_matrix.R
\name{validate_design_matrix}
\alias{validate_design_matrix}
\title{Validate design matrix}
\usage{
validate_design_matrix(design, T_ = NULL)
}
\arguments{
\item{design}{The design matrix}
}
\value{
The (modified) design matrix
}
\description{
Coerces \code{design} to a numeric matrix, and optionally checks that the
number of rows is as expected. Sets constant-valued columns to 1, and scales
all other columns.
}
\keyword{internal}
