% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_decomp.R
\name{var_decomp}
\alias{var_decomp}
\title{Compute variance decomposition}
\usage{
var_decomp(x, verbose = FALSE)
}
\arguments{
\item{x}{The data as a 3D array: measurements by subjects by variables.
(Alternatively, a matrix that is measurements by subjects, if only one
variable exists.)}

\item{verbose}{If \code{TRUE}, display progress of algorithm. Default:
\code{FALSE}.}
}
\value{
The variance decomposition.
}
\description{
Calculate the various ANOVA sums of squares for repeated measures data.
For each variable, subjects with any missing measurement are excluded.
}
