% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_regression.R
\name{check_design_matrix}
\alias{check_design_matrix}
\title{Check design matrix}
\usage{
check_design_matrix(design, T_ = nrow(design))
}
\arguments{
\item{design}{The design matrix}
}
\value{
The (modified) design matrix
}
\description{
Check design matrix
}
\keyword{internal}
