% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_FC.R
\name{color_palette}
\alias{color_palette}
\title{Palette}
\usage{
color_palette(pal = "Beach")
}
\arguments{
\item{pal}{\code{"Beach"} (default) is the only palette currently available.}
}
\value{
A data.frame with two columns: \code{"col"} has the hex code of color,
and \code{"val"} has the placement of colors between zero and one.
}
\description{
Color palette
}
\keyword{internal}
