% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3ddespike.R
\name{despike_3D}
\alias{despike_3D}
\title{despike_3D from AFNI}
\usage{
despike_3D(Yt, c1 = 2.5, c2 = 4)
}
\arguments{
\item{Yt}{The data vector.}

\item{c1}{spike threshold. Default: \code{2.5}.}

\item{c2}{upper range of the acceptable deviation from the fit. Default:
\code{4}.}
}
\description{
Identify and interpolate outliers.
}
