% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg.R
\name{dual_reg}
\alias{dual_reg}
\title{Dual Regression}
\usage{
dual_reg(
  BOLD,
  GICA,
  scale_by = c("mean", "sd", "none"),
  scale_sm_FWHM = 4,
  scale_sm_xifti = NULL,
  TR = NULL,
  hpf = 0.01,
  lpf = NULL,
  GSR = FALSE
)
}
\arguments{
\item{BOLD}{Subject-level fMRI data matrix (\eqn{V \times T}). Rows will be
centered.}

\item{GICA}{Group-level independent components (\eqn{V \times Q})}

\item{scale_by}{Scale the BOLD at each voxel based on either its
\code{"mean"} (default), its \code{"sd"}, or do not scale (\code{"none"}).
Mean scaling cannot be used if the \code{BOLD} have already been de-meaned.}

\item{scale_sm_FWHM}{Full width at half maximum (FWHM) for smoothing the
estimates of scale across brain locations (see \code{scale_by}), to reduce
the variance of the estimates. Set to \code{0} to disable smoothing, or
\code{Inf} for "global" smoothing (estimate and use one measure of scale
across the entire brain). Otherwise, for "local" smoothing, this should be a
positive number. Default: \code{4}.}

\item{scale_sm_xifti}{Required only for "local" smoothing of the scale
estimates with CIFTI data (see \code{scale_sm_FWHM}). To smooth the scale
estimates, provide a \code{"xifti"} object aligned with \code{BOLD}. If no
\code{"xifti"} object is provided (default) smoothing must be skipped.}

\item{TR}{The temporal resolution of the data, i.e. the time between volumes,
in seconds. \code{TR} is required for detrending with \code{hpf}.}

\item{hpf, lpf}{The frequencies at which to apply temporal filtering to the
data during pre-processing, in Hertz. Set either to \code{NULL} to disable.
Default: \code{0.01} Hz highpass filter, and \code{NULL} for the lowpass
filter (disabled). Filtering is accomplished by nuisance regression of
discrete cosine transform (DCT) bases.

The highpass filter serves to detrend the data, since low-frequency
variance is associated with noise. The lowpass filter removes high-frequency
variance, which is also thought to be from non-neuronal noise.

Note the \code{TR} argument is required for temporal filtering. If
\code{TR} is not provided, \code{hpf} and \code{lpf} will be ignored.}

\item{GSR}{Center BOLD across columns (each image)? This
is equivalent to performing global signal regression. Default:
\code{FALSE}.}
}
\value{
A list containing
the subject-level independent components \strong{S} (\eqn{Q \times V}),
and subject-level mixing matrix \strong{A} (\eqn{TxQ}).
}
\description{
Dual Regression
}
\examples{
nT <- 30
nV <- 400
nQ <- 7
mU <- matrix(rnorm(nV*nQ), nrow=nV)
mS <- mU \%*\% diag(seq(nQ, 1)) \%*\% matrix(rnorm(nQ*nT), nrow=nQ)
BOLD <- mS + rnorm(nV*nT, sd=.05) + 10
GICA <- mU
dual_reg(BOLD=BOLD, GICA=mU, scale_sm_FWHM=Inf, TR=.72)

}
