% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_reg_parc.R
\name{dual_reg_parc}
\alias{dual_reg_parc}
\title{Multiple regression for parcel data}
\usage{
dual_reg_parc(
  BOLD,
  parc,
  parc_vals,
  scale_by = c("mean", "sd", "none"),
  scale_sm_FWHM = 4,
  scale_sm_xifti = NULL,
  TR = NULL,
  hpf = 0.01,
  lpf = NULL,
  GSR = FALSE
)
}
\arguments{
\item{BOLD}{fMRI numeric data matrix (\eqn{V \times T})}

\item{parc}{The parcellation as an integer vector.}

\item{parc_vals}{The parcel values (keys) in desired order, e.g.
\code{sort(unique(parc))}.}

\item{scale_by}{Scale the BOLD at each voxel based on either its
\code{"mean"} (default), its \code{"sd"}, or do not scale (\code{"none"}).
Mean scaling cannot be used if the \code{BOLD} have already been de-meaned.}

\item{scale_sm_FWHM}{Full width at half maximum (FWHM) for smoothing the
estimates of scale across brain locations (see \code{scale_by}), to reduce
the variance of the estimates. Set to \code{0} to disable smoothing, or
\code{Inf} for "global" smoothing (estimate and use one measure of scale
across the entire brain). Otherwise, for "local" smoothing, this should be a
positive number. Default: \code{4}.}

\item{scale_sm_xifti}{Required only for "local" smoothing of the scale
estimates with CIFTI data (see \code{scale_sm_FWHM}). To smooth the scale
estimates, provide a \code{"xifti"} object aligned with \code{BOLD}. If no
\code{"xifti"} object is provided (default) smoothing must be skipped.}

\item{TR}{The temporal resolution of the data, i.e. the time between volumes,
in seconds. \code{TR} is required for detrending with \code{hpf}.}

\item{hpf, lpf}{The frequencies at which to apply temporal filtering to the
data during pre-processing, in Hertz. Set either to \code{NULL} to disable.
Default: \code{0.01} Hz highpass filter, and \code{NULL} for the lowpass
filter (disabled). Filtering is accomplished by nuisance regression of
discrete cosine transform (DCT) bases.

The highpass filter serves to detrend the data, since low-frequency
variance is associated with noise. The lowpass filter removes high-frequency
variance, which is also thought to be from non-neuronal noise.

Note the \code{TR} argument is required for temporal filtering. If
\code{TR} is not provided, \code{hpf} and \code{lpf} will be ignored.}

\item{GSR}{Center BOLD across columns (each image)? This
is equivalent to performing global signal regression. Default:
\code{FALSE}.}
}
\value{
A list containing
the subject-level independent components \strong{S} (\eqn{Q \times V}),
and subject-level mixing matrix \strong{A} (\eqn{TxQ}).
}
\description{
Multiple regression for parcel data
}
