% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fMRItools-package.R
\docType{package}
\name{fMRItools}
\alias{fMRItools-package}
\alias{fMRItools}
\title{fMRItools: Routines for Common fMRI Processing Tasks}
\description{
See \code{help(package="fMRItools")} for a list of functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mandymejia/fMRItools}
  \item Report bugs at \url{https://github.com/mandymejia/fMRItools/issues}
}

}
\author{
\strong{Maintainer}: Amanda Mejia \email{mandy.mejia@gmail.com}

Authors:
\itemize{
  \item Amanda Mejia \email{mandy.mejia@gmail.com}
  \item Damon Pham \email{damondpham@gmail.com} (\href{https://orcid.org/0000-0001-7563-4727}{ORCID})
}

Other contributors:
\itemize{
  \item Mark Fiecas \email{mfiecas@umn.edu} [contributor]
}

}
