% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{ICA-based harmonization}
\usage{
harmonize(
  BOLD,
  GICA,
  mask = NULL,
  gii_hemi = NULL,
  inds = NULL,
  scale = c("local", "global", "none"),
  scale_sm_surfL = NULL,
  scale_sm_surfR = NULL,
  scale_sm_FWHM = 2,
  TR = NULL,
  hpf = 0.01,
  GSR = FALSE,
  brainstructures = c("all"),
  verbose = TRUE,
  do_harmonize = FALSE
)
}
\arguments{
\item{BOLD}{Vector of subject-level fMRI data in one of the following
formats: CIFTI file paths, \code{"xifti"} objects, GIFTI file paths,
\code{"gifti"} objects, NIFTI file paths, \code{"nifti"} objects,
or \eqn{V \times T} numeric matrices, where \eqn{V} is the number of data
locations and \eqn{T} is the number of timepoints. For GIFTI or
\code{"gifti"} input, each entry can also be a length-two list where the
first entry is the left cortex and the second is the right cortex.}

\item{GICA}{Group ICA maps in a format compatible with \code{BOLD}. Can also
be a (vectorized) numeric matrix (\eqn{V \times Q}) no matter the format of
\code{BOLD}. Its columns will be centered.}

\item{mask}{Required if the entries of \code{BOLD} are NIFTI
file paths or \code{"nifti"} objects, optional for other formats. For
NIFTI, this is a brain map formatted as a binary array of the same spatial
dimensions as the fMRI data, with \code{TRUE} corresponding to in-mask
voxels. For other formats, a logical vector.}

\item{gii_hemi}{Which hemisphere, if BOLD is a single list of GIFTI data?
Should be \code{"left"} or \code{"right"}. If \code{NULL} (default), try to
infer from the metadata.}

\item{inds}{Numeric indices of the group ICs to include in the template. If
\code{NULL}, use all group ICs (default).

If \code{inds} is provided, the ICs not included will be removed after
calculating dual regression, not before. This is because removing the ICs
prior to dual regression would leave unmodelled signals in the data, which
could bias the templates.}

\item{scale}{\code{"global"} (default), \code{"local"}, or \code{"none"}.
Global scaling will divide the entire data matrix by the mean image standard
deviation (\code{mean(sqrt(rowVars(BOLD)))}). Local scaling will divide each
data location's time series by its estimated standard deviation.}

\item{scale_sm_surfL, scale_sm_surfR, scale_sm_FWHM}{Only applies if
\code{scale=="local"} and \code{BOLD} represents surface data (CIFTI or
GIFTI). To smooth the standard deviation estimates used for local scaling,
provide the surface geometries along which to smooth as GIFTI geometry files
or \code{"surf"} objects, as well as the smoothing FWHM (default: \code{2}).

If \code{scale_sm_FWHM==0}, no smoothing of the local standard deviation
estimates will be performed.

If \code{scale_sm_FWHM>0} but \code{scale_sm_surfL} and
\code{scale_sm_surfR} are not provided, the default inflated surfaces from
the HCP will be used.

To create a \code{"surf"} object from data, see
\code{\link[ciftiTools]{make_surf}}. The surfaces must be in the same
resolution as the \code{BOLD} data.}

\item{TR}{The temporal resolution of the data, i.e. the time between volumes,
in seconds. \code{TR} is required for detrending with \code{hpf}.}

\item{hpf}{The frequency at which to apply a highpass filter to the data
during pre-processing, in Hertz. Default: \code{0.01} Hertz. Set to \code{0}
to disable the highpass filter.

The highpass filter serves to detrend the data, since low-frequency
variance is associated with noise. Highpass filtering is accomplished by
nuisance regression of discrete cosine transform (DCT) bases.

Note the \code{TR} argument is required for highpass filtering. If
\code{TR} is not provided, \code{hpf} will be ignored.}

\item{GSR}{Center BOLD across columns (each image)? This
is equivalent to performing global signal regression. Default:
\code{FALSE}.}

\item{brainstructures}{Only applies if the entries of \code{BOLD} are CIFTI
file paths. This is a character vector indicating which brain structure(s)
to obtain: \code{"left"} (left cortical surface), \code{"right"} (right
cortical surface) and/or \code{"subcortical"} (subcortical and cerebellar
gray matter). Can also be \code{"all"} (obtain all three brain structures).
Default: \code{c("all")}.}

\item{verbose}{Display progress updates? Default: \code{TRUE}.}

\item{do_harmonize}{Perform the harmonization?  If not, will only extract and
return features to be harmonized.}
}
\description{
Harmonize data using ICA.
}
\details{
Let \eqn{V} be the number of data locations; \eqn{Q} be the number of group
ICs, and \eqn{N} be the number of fMRI sessions.
}
\keyword{internal}
