% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_format_ifti.R
\name{infer_format_ifti}
\alias{infer_format_ifti}
\title{Infer fMRI data format}
\usage{
infer_format_ifti(BOLD, verbose = FALSE)
}
\arguments{
\item{BOLD}{The fMRI data}

\item{verbose}{Print the format? Default: \code{FALSE}.}
}
\value{
A length-two vector. The first element indicates the format:
\code{"CIFTI"} file path, \code{"xifti"} object,
\code{"GIFTI"} file path, \code{"gifti"} object,
\code{"NIFTI"} file path, \code{"nifti"} object,
\code{"RDS"} file path, or \code{"data"}.
The second element indicates
the sub-format if relevant; i.e. the type of CIFTI or GIFTI file/object.
}
\description{
Infer fMRI data format
}
