% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readin.R, R/infer_format_ifti.R
\name{infer_format_ifti}
\alias{infer_format_ifti}
\title{Infer fMRI data format}
\usage{
infer_format_ifti(BOLD, verbose = FALSE)

infer_format_ifti(BOLD, verbose = FALSE)
}
\arguments{
\item{BOLD}{The fMRI data}

\item{verbose}{Print the format? Default: \code{FALSE}.}
}
\value{
The format: \code{"CIFTI"} file path, \code{"xifti"} object,
\code{"GIFTI"} file path, \code{"gifti"} object,
\code{"NIFTI"} file path, \code{"nifti"} object,
or \code{"data"}.

The format: \code{"CIFTI"} file path, \code{"xifti"} object,
\code{"NIFTI"} file path, \code{"nifti"} object, or \code{"data"}.
}
\description{
Infer fMRI data format

Infer fMRI data format
}
\keyword{Make}
\keyword{a}
\keyword{and}
\keyword{as}
\keyword{blah}
\keyword{data}
\keyword{formats}
\keyword{function}
\keyword{internal}
\keyword{list}
\keyword{that}
\keyword{the}
\keyword{wrapper}
