% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_BOLD.R
\name{norm_BOLD}
\alias{norm_BOLD}
\title{Normalize BOLD data}
\usage{
norm_BOLD(
  BOLD,
  drop_first = 0,
  nuisance = NULL,
  scrub = NULL,
  TR = NULL,
  hpf = 0.01,
  lpf = NULL,
  center_rows = TRUE,
  center_cols = FALSE,
  scale_by = c("mean", "sd", "FUN", "none"),
  scale_sm_FWHM = 4,
  scale_sm_xifti = NULL,
  scale_sm_xifti_mask = NULL,
  scale_FUN = NULL,
  give_stats = FALSE
)
}
\arguments{
\item{BOLD}{fMRI numeric data matrix (\eqn{V \times T})}

\item{drop_first}{(Optional) Number of volumes to drop from the start of each
BOLD session. Default: \code{0}.}

\item{nuisance}{(Optional) Nuisance matrix to regress from the BOLD data.

Nuisance regression is performed in a simultaneous regression with any spike
regressors from \code{scrub} and DCT bases from \code{hpf}.

Note that the nuisance matrix should be provided with timepoints matching
the original \code{BOLD} irregardless of \code{drop_first}.
Nuisance matrices will be truncated automatically if \code{drop_first>0}.}

\item{scrub}{(Optional) Numeric vector of integers giving the indices
of volumes to scrub from the BOLD data. (List the volumes to remove, not the
ones to keep.)

Scrubbing is performed within a nuisance regression by adding a spike
regressor to the nuisance design matrix for each volume to scrub.

Note that indices are counted beginning with the first index in the
\code{BOLD} session irregardless of \code{drop_first}. The indices will be
adjusted automatically if \code{drop_first>0}.}

\item{TR}{The temporal resolution of the data, i.e. the time between volumes,
in seconds. \code{TR} is required for detrending with \code{hpf}.}

\item{hpf, lpf}{The frequencies at which to apply temporal filtering to the
data during pre-processing, in Hertz. Set either to \code{NULL} to disable.
Default: \code{0.01} Hz highpass filter, and \code{NULL} for the lowpass
filter (disabled). Highpass filtering is accomplished by nuisance regression
of discrete cosine transform (DCT) bases. Lowpass filtering is accomplished
by a non-linear filter: \code{\link{fsl_bptf}}.

Note the \code{TR} argument is required for temporal filtering. If
\code{TR} is not provided, an error is raised if \code{hpf} and \code{lpf}
have been set.}

\item{center_rows, center_cols}{Center BOLD data across rows (each data
location's time series) or columns (each time point's image)? Default:
\code{TRUE} for row centering, and \code{FALSE} for column centering.}

\item{scale_by}{Scale the BOLD at each voxel based on either its
\code{"mean"} (default), its \code{"sd"}, or do not scale (\code{"none"}).
Mean scaling cannot be used if the \code{BOLD} have already been de-meaned.}

\item{scale_sm_FWHM}{Full width at half maximum (FWHM) for smoothing the
estimates of scale across brain locations (see \code{scale_by}), to reduce
the variance of the estimates. Set to \code{0} to disable smoothing, or
\code{Inf} for "global" smoothing (estimate and use one measure of scale
across the entire brain). Otherwise, for "local" smoothing, this should be a
positive number. Default: \code{4}.}

\item{scale_sm_xifti}{Required only for "local" smoothing of the scale
estimates with CIFTI data (see \code{scale_sm_FWHM}). To smooth the scale
estimates, provide a \code{"xifti"} object aligned with \code{BOLD}. If no
\code{"xifti"} object is provided (default) smoothing must be skipped.}

\item{scale_sm_xifti_mask}{For local smoothing of scale estimates, the data
must be unmasked to be mapped back to the surface. So if the data are
masked, provide the mask here.}

\item{scale_FUN}{If \code{scale_by} is \code{"FUN"}, this function is used
to compute the scaling measure from BOLD, after it has been centered
(and so after dropping any volumes, and applying nuisance regression,
scrubbing, and temporal filtering). Otherwise, ignored.}

\item{give_stats}{Return the intercept, residual SD estimates, and full
design matrix from the nuisance regression? Default: \code{FALSE}}
}
\value{
Normalized BOLD data matrix (\eqn{V \times T}), or if \code{give_stats},
a list with three elements: the normed BOLD, the intercept estimate, and the
residual SD estimate.
}
\description{
Clean the data (scrubbing, detrending, nuisance regression, and temporal
filtering in a simultaneous framework), then center it, then scale it, in
that order.
}
