% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_regression.R
\name{nuisance_regression}
\alias{nuisance_regression}
\title{Nuisance regression}
\usage{
nuisance_regression(Y, design)
}
\arguments{
\item{Y}{The \eqn{T} by \eqn{V} or \eqn{V} by \eqn{T} data.}

\item{design}{The \eqn{T} by \eqn{Q} matrix of nuisance regressors.}
}
\value{
The data after nuisance regression.
}
\description{
Performs nuisance regression. The data and design matrix must both be
centered, or an intercept must be included in the design matrix!
}
