% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_med.R
\name{scale_med}
\alias{scale_med}
\title{Robust scaling}
\usage{
scale_med(mat, TOL = 1e-08, drop_const = TRUE)
}
\arguments{
\item{mat}{A numerical matrix.}

\item{TOL}{minimum MAD to consider a column non-constant.
Default: \code{1e-8}}

\item{drop_const}{Drop}
}
\value{
The input matrix with its columns centered and scaled.
}
\description{
Centers and scales the columns of a matrix robustly
}
\details{
Centers each column on its median, and scales each column by its median
absolute deviation (MAD). Constant-valued columns are set to \code{NA}
(or removed if \code{drop_const}) and a warning is raised. If all
MADs are zero, an error is raised.
}
