% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UT.R
\name{UT2mat}
\alias{UT2mat}
\title{Upper Triangular Vector to Matrix}
\usage{
UT2mat(x, diag = 0, LT = 0)
}
\arguments{
\item{x}{A vector containing the upper triangular elements of a square,
symmetric matrix.}

\item{diag}{A scalar value to use for the diagonal values of the matrix, or
\code{"x"} if \code{x} includes the diagonal values. Default: \code{NA}.}

\item{LT}{A scalar value to use for the lower triangular values of the
matrix. Default: \code{0}.}
}
\value{
A symmetric matrix with the values of \code{x} in the upper and
lower triangles and the value \code{diag} on the diagonal.
}
\description{
Returns the symmatric square matrix from a vector containing
the upper triangular elements
}
